Push ProfileController methods down to ProfileService
